Fix Windows library search
As the `TBB_LIB_PATH_SUFFIX` is a semicolon seperated list on Windows, the double quotes have to be removed when expanding the variable. Else, the call to `find_library` will search for the single concatenated `PATH_SUFFIXES` and will not find anything at all.
PS: For me it works on Windows with this change (see your issue #1).